fix: widen NNTree Scalar from i32 to i64 to prevent rstar AABB overflow
rstar multiplies AABB dimension spans internally — overflows i32 when
entities span large map regions. All locate_within_distance call sites
updated to pass i64 squared distances.

+++ crates/server/src/systems/aoi.rs
@@ -28,11 +28,11 @@ use crate::{
 /// AOI radius: entities within this distance are visible to players.
 /// Covers FOV_CHUNK_RADIUS + 1 buffer chunk in all directions (hex chunks).
 pub const AOI_RADIUS: i32 = (FOV_CHUNK_RADIUS as i32 + 1) * CHUNK_SPACING as i32 + CHUNK_RADIUS as i32;
-const AOI_RADIUS_SQ: i32 = AOI_RADIUS * AOI_RADIUS;
+const AOI_RADIUS_SQ: i64 = AOI_RADIUS as i64 * AOI_RADIUS as i64;
 
 /// Exit radius: hysteresis buffer to prevent enter/exit flicker at the boundary.
 const EXIT_RADIUS: i32 = AOI_RADIUS + CHUNK_SPACING as i32;
-const EXIT_RADIUS_SQ: i32 = EXIT_RADIUS * EXIT_RADIUS;
+const EXIT_RADIUS_SQ: i64 = EXIT_RADIUS as i64 * EXIT_RADIUS as i64;
 
 /// Updates LoadedBy membership when entities move.
 ///
