+++ crates/server/src/systems/behaviour/chase.rs
@@ -291,7 +291,7 @@ pub fn chase(
                 // Close enough to spawn - search for new target
                 let nearby = nntree.locate_within_distance(
                     *npc_loc,
-                    chase_config.acquisition_range as i32 * chase_config.acquisition_range as i32,
+                    chase_config.acquisition_range as i64 * chase_config.acquisition_range as i64,
                 );
 
                 let valid_targets: Vec<Entity> = nearby
